requirements.txt
# ═══════════════════════════════════════════════════════════════════════════════
#                    متطلبات تشغيل بوت التداول
# ═══════════════════════════════════════════════════════════════════════════════

# مكتبة Binance الرسمية
python-binance==1.0.19

# مكتبة التحليل الفني
numpy>=1.24.0

# لإرسال طلبات HTTP
requests>=2.28.0

# للتعامل مع الوقت والتواريخ
python-dateutil>=2.8.0

# لخادم الويب (Render)
flask>=3.0.0
